+++ applications/fishing-map/src/features/map/popups/FishingLayers.tsx
@@ -11,7 +11,11 @@ import {
   getRelatedDatasetsByType,
 } from 'features/datasets/datasets.selectors'
 import I18nNumber from 'features/i18n/i18nNumber'
-import { TooltipEventFeature, useClickedEventConnect } from 'features/map/map.hooks'
+import {
+  SUBLAYER_INTERACTION_TYPES_WITH_VESSEL_INTERACTION,
+  TooltipEventFeature,
+  useClickedEventConnect,
+} from 'features/map/map.hooks'
 import { formatI18nDate } from 'features/i18n/i18nDate'
 import { ExtendedFeatureVessel } from 'features/map/map.slice'
 import { isUserLogged } from 'features/user/user.selectors'
@@ -116,8 +120,12 @@ function FishingTooltipRow({ feature, showFeaturesDetails }: FishingTooltipRowPr
               const vesselLabel = getVesselLabel(vessel, true)
               return (
                 <button
-                  disabled={feature.category === 'presence'}
                   key={i}
+                  disabled={
+                    !SUBLAYER_INTERACTION_TYPES_WITH_VESSEL_INTERACTION.includes(
+                      feature.temporalgrid?.sublayerInteractionType || ''
+                    )
+                  }
                   className={styles.vesselRow}
                   onClick={() => onVesselClick(vessel)}
                 >