+++ applications/fishing-map/src/features/map/popups/PopupWrapper.tsx
@@ -67,8 +67,17 @@ function PopupWrapper({
                 />
               ))
             case DataviewCategory.Presence:
-              return features.map((feature, i) =>
-                feature.temporalgrid?.sublayerInteractionType === 'viirs' ? (
+              return features.map((feature, i) => {
+                if (feature.temporalgrid?.sublayerInteractionType === 'presence-detail') {
+                  return (
+                    <FishingTooltipRow
+                      key={i + (feature.title as string)}
+                      feature={feature}
+                      showFeaturesDetails={type === 'click'}
+                    />
+                  )
+                }
+                return feature.temporalgrid?.sublayerInteractionType === 'viirs' ? (
                   <ViirsTooltipRow
                     key={i + (feature.title as string)}
                     feature={feature}
@@ -81,7 +90,7 @@ function PopupWrapper({
                     showFeaturesDetails={type === 'click'}
                   />
                 )
-              )
+              })
             case DataviewCategory.Events:
               return (
                 <TileClusterRow