+++ packages/dataviews-client/src/resolve-dataviews-generators.ts
@@ -335,12 +335,20 @@ export function getDataviewsGeneratorConfigs(
       }
       const interactionTypes = uniq(
         dataview.datasets?.map((dataset) => dataset.configuration?.type || 'fishing-effort')
-      )
+      ) as HeatmapAnimatedInteractionType[]
       if (interactionTypes.length !== 1) {
         throw new Error(
           `Shouldnt have distinct dataset config types for the same heatmap layer: ${interactionTypes.toString()}`
         )
       }
+      const interactionType =
+        // Some VMS presence layers have interaction, this is the way of
+        // allowing it but keeping it disabled in the global one
+        interactionTypes[0] === 'presence'
+          ? dataview?.config?.presenceDetails === true
+            ? 'presence-detail'
+            : 'presence'
+          : (interactionTypes[0] as HeatmapAnimatedInteractionType)
       const sublayer: HeatmapAnimatedGeneratorSublayer = {
         id: dataview.id,
         datasets,
@@ -353,7 +361,7 @@ export function getDataviewsGeneratorConfigs(
           unit: units[0],
           color: dataview?.config?.color,
         },
-        interactionType: interactionTypes[0] as HeatmapAnimatedInteractionType, // TODO I don't understand why dataset.configuration?.type is of type EventTypes?
+        interactionType,
       }
 
       return sublayer