+++ applications/dataviews-editor/CHANGELOG.md
@@ -4,9 +4,9 @@
 
 ### Patch Changes
 
-- Updated dependencies [c60ce91a]
-  - @globalfishingwatch/dataviews-client@11.4.1
-  - @globalfishingwatch/layer-composer@8.5.1
+- Updated dependencies [3513a689]
+  - @globalfishingwatch/dataviews-client@11.5.0
+  - @globalfishingwatch/layer-composer@8.6.0
 
 ## 1.1.68